Move lexer into separate files as discussed in #10.
